+++ NPCMapLocations/docs/release-notes.md
@@ -6,6 +6,7 @@
 * Fixed error when a location has invalid warp targets.
 * Fixed map positioning on the beach farm (thanks to RushedLogic!).
 * Fixed misaligned minimap texture on the beach farm (thanks to RushedLogic!).
+* Fixed villager names not translated in map tooltips.
 * Internal refactoring and optimizations.
 
 ## 2.6.2